Add profile styling
