+++ src/components/Mission.scss
@@ -1,3 +1,28 @@
 h1 {
   text-align: center;
 }
+table {
+  width: 100%;
+  border-collapse: collapse;
+}
+tr {
+  width: 100%;
+}
+
+h3 {
+  width: max-content;
+  font-weight: 500;
+}
+
+td,
+th {
+  border: 1px solid #ddd;
+  text-align: left;
+  padding: 8px;
+  line-height: 26px;
+}
+
+th {
+  font-weight: 500;
+  font-size: 1.1rem;
+}
