+++ src/components/MyProfile.js
@@ -1,4 +1,5 @@
 import { useSelector } from 'react-redux';
+import './profile.scss';
 
 const MyProfile = () => {
   const missionsL = useSelector((state) => state.missions);
@@ -6,8 +7,8 @@ const MyProfile = () => {
   const rocketsL = useSelector((state) => state.rockets);
   const newRockets = rocketsL.filter((rocket) => rocket.canceled);
   return (
-    <div>
-      <div>
+    <div className="profile-section">
+      <div className="mission-div">
         <h2>My Missions</h2>
         <ul>
           {newMissions.length >= 1 ? (
@@ -21,7 +22,7 @@ const MyProfile = () => {
           )}
         </ul>
       </div>
-      <div>
+      <div className="mission-div">
         <h2>My Rockets</h2>
         <ul>
           {newRockets.length >= 1 ? (
